requirements.txt
cffi
cryptography
pycparser
PyMySQL
python-dotenv
six
